Replace helper scripts with Python
The helper scripts were tied to 'cog.yaml', so we couldn't put
them in the base image. This meant prediction required all the
code in the image.

Also, this is much neater and requires less string concatenation.

Also, the model option is no longer required to build a full
image.

Signed-off-by: Ben Firshman <ben@firshman.co.uk>