Set all labels after build
A quote in a `pre_install` step was being escaped incorrectly.
Unsurprisingly -- we were putting a shell script inside an argument
inside a string, inside a JSON document, inside a string, inside a
Dockerfile instruction.

By doing this in the post-build labelling step this works properly
(shrug). We should be really be using proper APIs as much as we can
or we're going to be endlessly battling typing/escaping problems
like this.

Signed-off-by: Ben Firshman <ben@firshman.co.uk>