Rename success status to succeeded
Adopting …ing and …ed endings for statuses will make it easier to add
consistent new statuses in the future. We currently have success and
failure, so we're going to either end up with inconsistent statuses
like, or we'll need to add a status like `cancellation` which feels
super weird.

With …ing and …ed, if we want to add statuses for jobs that haven't
started or are in progress, it's easy to find verbs: queuing, waiting,
running.

Take the opportunity to use the `Status` constant in the queue API
rather than a string, to move the two APIs closer together in behaviour
and implementation.

Signed-off-by: Dominic Baggott <dominic.baggott@gmail.com>