Add option to exit redis worker after too many failures
We're occasionally seeing models get into a zombie state in Replicate,
where the model will keep accepting jobs but it fails every single one.
When it's failing them very quickly, and survives as the final worker
after the model is downscaled, we end up serving only failures for the
affected model.

This change allows a maximum consecutive number of failures to be set.
If the worker fails more than that many predictions in a row, it'll
exit, forcing a new (hopefully healthy) worker to be started to take its
place.

Signed-off-by: Dominic Baggott <dominic.baggott@gmail.com>