Request and wait for an explicit healthcheck before each prediction
We wait up to 10 seconds for an explicit confirmation of the model
container's health before fetching the next prediction from Redis. This
should hopefully prevent us from dequeueing predictions that are bound
to fail (e.g. when the model container has crashed but we have not yet
detected that with the regular background checks).

Signed-off-by: Nick Stenning <nick@whiteink.com>