Make PUT endpoint idempotent
This rearranges a few things in order to implement what is hopefully
reasonably reliable idempotency for the PUT endpoint.

The core of the change is making PredictionRunner.predict() itself
idempotent if (and only if) the prediction request includes an ID. If
the prediction request includes an ID, then subsequent calls to
predict() with the same ID will return the preliminary response and
async result object as normal.

The behaviour of the HTTP endpoints is as follows:

- Both sync and async POST endpoints will return a 409 Conflict
  immediately if the runner is busy, regardless of whether or not an ID
  is specified in the prediction request.

- A second synchronous PUT for the same prediction ID received while
  the prediction is running will wait for prediction completion just
  like the initial request.

- A second asynchronous PUT for the same prediction ID received while
  the prediction is running will return the initial state of the
  prediction just like the initial request.

- A second PUT (either asynchronous or synchronous) with a prediction ID
  other than the running prediction will receive a 409 Conflict.

Signed-off-by: Nick Stenning <nick@whiteink.com>