Set a minimum CUDA version for Tensorflow compatibility (#1715)
We fetch the Tensorflow compatibility matrix directly from
the tensorflow release page but we include all versions without
checking for a minimal CUDA version. That conflicts with other code
paths that expect a smaller set of CUDA versions (e.g. cuda_base_images.json).

The change here introduces a minimum CUDA version for TensorFlow when
generating the compatibility matrix.

We are not making the CUDA version configurable as we only use the matrix
from the config/compatibility package and we don't have a use case to modify 
the compatibility matrix without making a code change.

Signed-off-by: Marcus Martins <marcus@replicate.com>