Build python package separately
This splits the building of the Python package into its own job, and
then runs the tests against the built package. This gives us the
greatest chance of testing exactly the code we run in a consumer
environment.

It will also allow us to build and push the PyPI package in a separate
workflow, knowing that the package we build from the same source ref
will be identical.