Download Path inputs in background
PR #1865 introduced a regression: URLs for Path inputs were downloaded
inside the request handler for `POST/PUT /predictions`.

This creates at least two problems:

1. It blocks the uvicorn thread, stopping us from serving healthchecks.
2. It makes the prediction create request take too long, even when
   `Prefer: respond-async` is specified.

Both of these issues cause unexpected failures in production.

This commit moves responsibility for fetching these files into Worker,
and specifically into its background thread. This is probably not a
long-term fix, as we want to make Worker (and ChildWorker) async
eventually.

This also adds a test which fails on main and is fixed by this change.