Publish on tag push rather than release
Unfortunately the `release` event associated with the release created by
`goreleaser` in the CI workflow does not trigger workflow runs. This is
because the release is created using the `$GITHUB_TOKEN`, which doesn't
trigger workflows to eliminate the possibility of recursive workflow
triggers. See:

  https://docs.github.com/en/actions/security-for-github-actions/security-guides/automatic-token-authentication#using-the-github_token-in-a-workflow

This changes the publishing workflow (which is separate so that PyPI
trusted publishing can reference specifically this workflow, and not
allow publishing from any CI workflow run) to trigger on tag push, just
like the main CI workflow.

The drawback here is that this is not conditional on the tests running
and passing. For now, this is probably ok, as we should not be pushing
tags until we see that the relevant commit has passed CI.