Clarify default install directory (#1912)
The install script seems to suggest `/usr/local/bin` as the default install directory. However, pressing `RETURN` will actually set the install dir to an empty string. The error message is not helpful as it then reads `The directory  does not exist.` (note the additional space).

This fix makes the script behave as expected: pressing `RETURN` accepts the default suggestion.

Signed-off-by: André Knörig <aknoerig@users.noreply.github.com>