Add .python-version to .dockerignore
Fixes #2275

There is a bug where an existing .python-version file will be copied
from an existing project into the Docker container and conflict with the
Python version set in the cog.yaml.

We now include the .python-version in the .dockerignore file so that
it will not be copied into the container at build time.