Refactor PostNewPipeline to use Replicate public API (#2378)
This follows the procedure for shipping an experimental "pipeline" model
via the API used by other areas of the Replicate product.

We use the new cog.replicate.com/api/token/{account} endpoint to
exchange the cog token for a short lived Replicate API token that
is restricted only to read/write actions on model resources.